+++ app/cmd/buki.js
@@ -1,5 +1,5 @@
 const request = require("request");
-const common = require("./common.js");
+const common = require("../common.js");
 const weaponsUrl = "https://stat.ink/api/v2/weapon";
 
 const bukiTypes = {
