+++ app/cmd/omikuji.js
@@ -1,4 +1,4 @@
-const common = require("./common.js");
+const common = require("../common.js");
 const kujis = {
   0: "大吉",
   1: "吉",
