+++ app/handler.js
@@ -1,18 +1,18 @@
-const handleBan = require("./ban.js");
-const handleBuki = require("./buki.js");
+const handleBan = require("./cmd/admin-cmd/ban.js");
+const handleBuki = require("./cmd/buki.js");
 const handleFriendCode = require("./cmd/friendcode.js");
 const handleSpecial = require("./cmd/special.js");
-const handleHelp = require("./help.js");
+const handleHelp = require("./cmd/help.js");
 const handleKansen = require("./cmd/kansen.js");
 const handlePick = require("./cmd/pick.js");
-const handleOmikuji = require("./omikuji.js");
+const handleOmikuji = require("./cmd/omikuji.js");
 const handlePoll = require("./cmd/poll.js");
 const handleRecruit = require("./cmd/recruit.js");
-const handleRule = require("./rule.js");
-const handleShow = require("./show.js");
-const handleStageInfo = require("./stageinfo.js");
+const handleRule = require("./cmd/rule.js");
+const handleShow = require("./cmd/show.js");
+const handleStageInfo = require("./cmd/stageinfo.js");
 const handleSub = require("./cmd/sub.js");
-const handleTimer = require("./timer.js");
+const handleTimer = require("./cmd/timer.js");
 const handleVoicePick = require("./cmd/vpick.js");
 const handleWiki = require("./cmd/wiki.js");
