+++ app/index.js
@@ -1,17 +1,23 @@
 // Discord bot implements
-const Discord = require("discord.js");
-const client = new Discord.Client({
-  partials: ["MESSAGE", "CHANNEL", "REACTION"],
+const { Client, Intents } = require("discord.js");
+const client = new Client({
+  intents: [Intents.FLAGS.GUILDS,
+  Intents.FLAGS.GUILD_MESSAGES,
+  Intents.FLAGS.GUILD_BANS,
+  Intents.FLAGS.GUILD_PRESENCES,
+  Intents.FLAGS.GUILD_VOICE_STATES,
+  Intents.FLAGS.GUILD_MESSAGE_REACTIONS,
+  ],
 });
 const Handler = require("./handler.js");
 const Dispandar = require("./event/dispandar.js");
 const TTS = require("./tts/voice_bot_node.js");
 const privateChat = require("./voice/secretchat.js");
-const handleStageInfo = require("./stageinfo.js");
+const handleStageInfo = require("./cmd/stageinfo.js");
 const removeRookie = require("./event/rookie.js");
 const chatCountUp = require("./event/members.js");
 const suggestionBox = require("./reaction/suggestion-box.js");
-const oneHourLeague = require("./one_hour_league.js");
+const oneHourLeague = require("./cmd/one_hour_league.js");
 const join = require("./event/join.js");
 const deleteToken = require("./event/delete_token.js");
 client.login(process.env.DISCORD_BOT_TOKEN);
@@ -38,9 +44,9 @@ client.on("messageCreate", async (msg) => {
   if (msg.content.match("ボーリング")) {
     msg.channel.send(
       "```「ボウリング」とは、前方に正三角形に並べられた10本のピンと呼ばれる棒をめがけボールを転がし、倒れたピンの数によって得られる得点を競うスポーツでし。" +
-        "専用施設のボウリング場に設置された細長いレーンの上で行われる屋内競技で、レーンの長さが約23m、ピンまでの距離は約18mで行われるのが一般的でし。" +
-        "英語では “bowling” と書き、球を意味する “ball” ではなく、ラテン語で「泡」や「こぶ」を意味する “bowl” が語源とされているでし。" +
-        "\n文部科学省は国語審議会で、球技を指す場合は「ボウリング」表記を用い、掘削を意味する「ボーリング」と区別することを推奨しているでし。```"
+      "専用施設のボウリング場に設置された細長いレーンの上で行われる屋内競技で、レーンの長さが約23m、ピンまでの距離は約18mで行われるのが一般的でし。" +
+      "英語では “bowling” と書き、球を意味する “ball” ではなく、ラテン語で「泡」や「こぶ」を意味する “bowl” が語源とされているでし。" +
+      "\n文部科学省は国語審議会で、球技を指す場合は「ボウリング」表記を用い、掘削を意味する「ボーリング」と区別することを推奨しているでし。```"
     );
   }
