+++ app/voice/secretchat.js
@@ -8,7 +8,16 @@
 */
 // Discord bot implements
 const Discord = require("discord.js");
-const client = new Discord.Client();
+const { Client, Intents } = require("discord.js");
+const client = new Client({
+  intents: [Intents.FLAGS.GUILDS,
+  Intents.FLAGS.GUILD_MESSAGES,
+  Intents.FLAGS.GUILD_BANS,
+  Intents.FLAGS.GUILD_PRESENCES,
+  Intents.FLAGS.GUILD_VOICE_STATES,
+  Intents.FLAGS.GUILD_MESSAGE_REACTIONS,
+  ],
+});
 const request = require("request");
 
 module.exports = {
